Add base operator decorators
